크고 작은 UI 수정

+++ app/src/main/java/com/kims/recipe2/model/Ingredient.kt
@@ -9,7 +9,7 @@ data class Ingredient(
     val category: String = "",
     val location: String = "",   // 냉장고 위치 (예: "냉동실")
     val quantity: Int = 1,       // 개수
-    val amount: Double = 0.0,    // 양 (g, ml 등)
+    val amount: Int = 0,    // 양 (g, ml 등)
     val unit: String = "개",      // 단위 (개, g, ml 등)
     val expirationDate: Date? = null, // 유통기한 필드 추가
     // 👇 영양소 필드 추가
